requirements.txt
anyio==3.6.2
asgiref==3.5.2
click==8.1.3
colorama==0.4.6
fastapi==0.86.0
h11==0.14.0
idna==3.4
kbinxml==1.7
lxml==4.9.1
pycryptodomex==3.15.0
pydantic==1.10.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tinydb==4.7.0
typing_extensions==4.4.0
uvicorn==0.19.0
